requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0

# API clients
openai>=1.0.0
anthropic>=0.8.0

# Data processing
jsonlines>=3.0.0
pyyaml>=6.0
tqdm>=4.62.0

# Statistical analysis
scikit-learn>=1.0.0
statsmodels>=0.13.0

# Visualization
plotly>=5.0.0
kaleido>=0.2.1

# Text processing
nltk>=3.6.0
spacy>=3.4.0

# Utilities
python-dotenv>=0.19.0
click>=8.0.0
rich>=12.0.0

# Testing and quality
pytest>=6.2.0
black>=22.0.0
flake8>=4.0.0
